refactor: Inputparser ui 패키지로 이동
