+++ src/test/kotlin/lotto/InputParserTest.kt
@@ -1,5 +1,6 @@
 package lotto
 
+import lotto.ui.InputParser
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 import org.junit.jupiter.params.ParameterizedTest
